+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Properly reconstruct the original pool in the ClassConverter transformer if a class is found to be available. (remm)
 - Do not buffer very large STORED zip entries when processing them in streaming mode. (remm)
 - Various minor fixes from code review. (remm)
+- Improve logging handling for Ant tasks. (remm)
 
 ## 1.0.12
 - Add Maven Wrapper Plugin to manage the Maven wrapper. (markt)